Update of unit tests to use KRaft - part 5
